feature: fix add find all-find id Users
