+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export { default as playlistRouter } from "./playlist";
 export { default as songrouter } from "./songs";
+export { default as userrouter } from "./users";
+
