+++ src/components/songs/controller.ts
@@ -69,16 +69,3 @@ export const creacion= async(req:Request, res: Response): Promise<void> =>{
 }
 
 
-
-export const findAllUsers = async (_req: Request, res: Response): Promise<void> => {
-    try {
-        const users = await prisma.user.findMany();
-
-        res.status(200).json({
-            ok: true,
-            data: users,
-        });
-    } catch (error) {
-        res.status(500).json({ ok: false, message: error });
-    }
-}
